+++ packages/extensions/src/heading/types.ts
@@ -1,0 +1,4 @@
+
+export interface HeadingAttrs {
+  level: number;
+}
